Ripped out GameController from App
